pyproject.toml
[tool.poetry]
name = "active-user-checker-discord"
version = "0.1.0"
description = ""
authors = ["Mikhail Moskovkin <anstane228@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.3.2"
python-dotenv = "^1.0.0"
aiogram = "^3.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
